modules finish

+++ Backend/public/controllers/controller.Event.js
@@ -15,14 +15,14 @@ import CustomError from "../components/CustomError";
 import logger from "../components/logger";
 const mgClient = new MongoClient(AppConfig.mongoURL);
 export class EventController {
-    static create(userId, eventData) {
+    static create(ownerId, eventData) {
         return __awaiter(this, void 0, void 0, function* () {
             return yield errorHandlerController(() => __awaiter(this, void 0, void 0, function* () {
                 yield mgClient.connect();
                 const db = mgClient.db("Notebook");
                 const newEvent = {
                     id: generateID(),
-                    userId: userId,
+                    ownerId: ownerId,
                     title: eventData.title,
                     description: eventData.decription,
                     date: eventData.date,
@@ -36,7 +36,7 @@ export class EventController {
             }));
         });
     }
-    static remove(userId, eventId) {
+    static remove(ownerId, eventId) {
         return __awaiter(this, void 0, void 0, function* () {
             return yield errorHandlerController(() => __awaiter(this, void 0, void 0, function* () {
                 yield mgClient.connect();
@@ -46,13 +46,18 @@ export class EventController {
                     logger.debug("EventController.remove -> Event don`t found");
                     throw new CustomError("DATA_DONT_FOUND", 404, "Event don`t found");
                 }
+                if (eventData.ownerId != ownerId) {
+                    let message = "Forbidden";
+                    logger.debug(message);
+                    throw new CustomError("FORBIDDEN", 403, message);
+                }
                 yield db.collection("Events").deleteOne({ id: eventId });
                 mgClient.close();
                 logger.info("EventController.remove -> OK");
             }));
         });
     }
-    static chamgeData(eventId, newEventData) {
+    static changeData(ownerId, eventId, newEventData) {
         return __awaiter(this, void 0, void 0, function* () {
             return yield errorHandlerController(() => __awaiter(this, void 0, void 0, function* () {
                 yield mgClient.connect();
@@ -62,13 +67,18 @@ export class EventController {
                     logger.debug("EventController.chamgeData -> Event don`t found");
                     throw new CustomError("DATA_DONT_FOUND", 404, "Event don`t found");
                 }
+                if (eventData.ownerId != ownerId) {
+                    let message = "Forbidden";
+                    logger.debug(message);
+                    throw new CustomError("FORBIDDEN", 403, message);
+                }
                 yield db.collection("Events").updateOne({ _id: eventData._id }, newEventData);
                 mgClient.close();
                 logger.info("EventController.chamgeData -> OK");
             }));
         });
     }
-    static getData(eventId) {
+    static getData(ownerId, eventId) {
         return __awaiter(this, void 0, void 0, function* () {
             return yield errorHandlerController(() => __awaiter(this, void 0, void 0, function* () {
                 yield mgClient.connect();
@@ -78,13 +88,17 @@ export class EventController {
                     logger.debug("EventController.getData -> Event don`t found");
                     throw new CustomError("DATA_DONT_FOUND", 404, "Event don`t found");
                 }
+                if (!ownerId == eventData.ownerId) {
+                    logger.debug("NoteController.removeById -> FORIBBEN", 403, "There is no access");
+                    throw new CustomError("FORIBBEN", 403, "There is no access");
+                }
                 mgClient.close();
                 logger.info("EventController.getData -> OK");
                 return eventData;
             }));
         });
     }
-    static getList(eventId, listType) {
+    static getList(ownerId, eventId, listType) {
         return __awaiter(this, void 0, void 0, function* () {
             return yield errorHandlerController(() => __awaiter(this, void 0, void 0, function* () {
                 yield mgClient.connect();
@@ -93,6 +107,11 @@ export class EventController {
                 if (!eventData) {
                     logger.debug("EventController.getList -> Event don`t found");
                     throw new CustomError("DATA_DONT_FOUND", 404, "Event don`t found");
+                }
+                if (eventData.ownerId != ownerId) {
+                    let message = "Forbidden";
+                    logger.debug(message);
+                    throw new CustomError("FORBIDDEN", 403, message);
                 }
                 mgClient.close();
                 logger.info("EventController.getList -> OK");
