+++ Backend/public/controllers/controller.Note.js
@@ -14,18 +14,22 @@ import { AppConfig } from "../config";
 import { generateID } from "../components/generator";
 const mgClient = new MongoClient(AppConfig.mongoURL);
 export class NoteController {
-    static get(noteId, userId) {
-        return __awaiter(this, void 0, void 0, function* () {
+    static get(noteId_1, ownerId_1) {
+        return __awaiter(this, arguments, void 0, function* (noteId, ownerId, collection = "Users") {
             try {
                 yield mgClient.connect();
                 const db = mgClient.db("Notebook");
+                const idnames = {
+                    Users: "ownerId",
+                    Schedules: "scheduleId"
+                };
                 const noteData = yield db.collection("Notes").findOne({ id: noteId });
                 if (!noteData) {
                     let message = "Note dot`t found";
                     logger.debug(message);
                     throw new CustomError("DATA_DONT_EXISIT", 404, message);
                 }
-                if (noteData.userId != userId) {
+                if (noteData.ownerId != ownerId) {
                     let message = "Forbidden";
                     logger.debug(message);
                     throw new CustomError("FORBIDDEN", 403, message);
@@ -46,12 +50,12 @@ export class NoteController {
             }
         });
     }
-    static create(userId, note) {
+    static create(ownerId, note) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
                 const noteData = {
                     id: generateID(),
-                    userId: userId,
+                    ownerId: ownerId,
                     title: note.title,
                     content: note.content
                 };
@@ -74,12 +78,12 @@ export class NoteController {
             }
         });
     }
-    static getAll(userId_1) {
-        return __awaiter(this, arguments, void 0, function* (userId, collection = "Users") {
+    static getAll(ownerId_1) {
+        return __awaiter(this, arguments, void 0, function* (ownerId, collection = "Users") {
             try {
                 yield mgClient.connect();
                 const db = mgClient.db("Notebook");
-                const arrayNotesId = (yield db.collection(collection).findOne({ id: userId })).notes;
+                const arrayNotesId = (yield db.collection(collection).findOne({ id: ownerId })).notes;
                 const arrayNotes = yield db.collection("Notes").find({ $elemMatch: { id: { $in: arrayNotesId } } });
                 if (!arrayNotes) {
                     return [];
@@ -100,7 +104,7 @@ export class NoteController {
             }
         });
     }
-    static removeById(noteId) {
+    static removeById(ownerId, noteId) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
                 yield mgClient.connect();
@@ -110,6 +114,10 @@ export class NoteController {
                     let message = "Note dot`t found";
                     logger.debug(message);
                     throw new CustomError("DATA_DONT_EXISIT", 404, message);
+                }
+                if (!ownerId == note.ownerId) {
+                    logger.debug("NoteController.removeById -> FORIBBEN", 403, "There is no access");
+                    throw new CustomError("FORIBBEN", 403, "There is no access");
                 }
                 yield db.collection("Note").deleteOne({ _id: note._id });
                 return;
@@ -128,7 +136,7 @@ export class NoteController {
             }
         });
     }
-    static changeContent(noteId, newContent) {
+    static changeContent(ownerId, noteId, newContent) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
                 yield mgClient.connect();
@@ -138,6 +146,10 @@ export class NoteController {
                     let message = "Note dot`t found";
                     logger.debug(message);
                     throw new CustomError("DATA_DONT_EXISIT", 404, message);
+                }
+                if (!ownerId == noteData.ownerId) {
+                    logger.debug("NoteController.removeById -> FORIBBEN", 403, "There is no access");
+                    throw new CustomError("FORIBBEN", 403, "There is no access");
                 }
                 yield db.collection("Notes").updateOne({ _id: noteData._id }, { content: newContent });
                 return;
