+++ Backend/src/controllers/controller.Note.ts
@@ -8,10 +8,15 @@ import { TNoteContent, TNoteData } from "../types";
 const mgClient = new MongoClient(AppConfig.mongoURL)
 
 export class NoteController {
-    static async get(noteId: string, userId: string): Promise<WithId<Document>> {
+    static async get(noteId: string, ownerId: string, collection: "Users" | "Schedules" = "Users"): Promise<WithId<Document>> {
         try {
             await mgClient.connect()
             const db = mgClient.db("Notebook")
+
+            const idnames = {
+                Users: "ownerId",
+                Schedules: "scheduleId"
+            }
 
             const noteData = await db.collection("Notes").findOne({ id: noteId })
             if (!noteData) {
@@ -19,7 +24,7 @@ export class NoteController {
                 logger.debug(message)
                 throw new CustomError("DATA_DONT_EXISIT", 404, message)
             }
-            if (noteData.userId != userId) {
+            if (noteData.ownerId != ownerId) {
                 let message = "Forbidden"
                 logger.debug(message)
                 throw new CustomError("FORBIDDEN", 403, message)
@@ -40,11 +45,11 @@ export class NoteController {
         }
     }
 
-    static async create(userId: string, note: TNoteData): Promise<void> {
+    static async create(ownerId: string, note: TNoteData): Promise<void> {
         try {
             const noteData: TNoteData = {
                 id: generateID(),
-                userId: userId,
+                ownerId: ownerId,
                 title: note.title,
                 content: note.content
             }
@@ -69,12 +74,12 @@ export class NoteController {
         }
     }
 
-    static async getAll(userId: string, collection:"Users"|"Schedules"="Users"): Promise<FindCursor<WithId<Document>> | Array<any>> {
+    static async getAll(ownerId: string, collection: "Users" | "Schedules" = "Users"): Promise<FindCursor<WithId<Document>> | Array<any>> {
         try {
             await mgClient.connect()
             const db = mgClient.db("Notebook")
 
-            const arrayNotesId = (await db.collection(collection).findOne({ id: userId })).notes
+            const arrayNotesId = (await db.collection(collection).findOne({ id: ownerId })).notes
             const arrayNotes = await db.collection("Notes").find({ $elemMatch: { id: { $in: arrayNotesId } } })
             if (!arrayNotes) {
                 return []
@@ -96,7 +101,7 @@ export class NoteController {
         }
     }
 
-    static async removeById(noteId: string): Promise<void> {
+    static async removeById(ownerId: string, noteId: string): Promise<void> {
         try {
             await mgClient.connect()
             const db = mgClient.db("Notebook")
@@ -106,6 +111,10 @@ export class NoteController {
                 let message = "Note dot`t found"
                 logger.debug(message)
                 throw new CustomError("DATA_DONT_EXISIT", 404, message)
+            }
+            if (!ownerId == note.ownerId) {
+                logger.debug("NoteController.removeById -> FORIBBEN", 403, "There is no access")
+                throw new CustomError("FORIBBEN", 403, "There is no access")
             }
             await db.collection("Note").deleteOne({ _id: note._id })
             return
@@ -121,21 +130,25 @@ export class NoteController {
         }
         finally {
             mgClient.close()
-        } 
+        }
     }
 
-    static async changeContent(noteId: string, newContent: TNoteContent): Promise<void> {
+    static async changeContent(ownerId:string, noteId: string, newContent: TNoteContent): Promise<void> {
         try {
             await mgClient.connect()
             const db = mgClient.db("Notebook")
 
-            const noteData = await db.collection("Notes").findOne({id:noteId})
+            const noteData = await db.collection("Notes").findOne({ id: noteId })
             if (!noteData) {
                 let message = "Note dot`t found"
                 logger.debug(message)
                 throw new CustomError("DATA_DONT_EXISIT", 404, message)
             }
-            await db.collection("Notes").updateOne({_id:noteData._id}, {content:newContent})
+            if (!ownerId == noteData.ownerId) {
+                logger.debug("NoteController.removeById -> FORIBBEN", 403, "There is no access")
+                throw new CustomError("FORIBBEN", 403, "There is no access")
+            }
+            await db.collection("Notes").updateOne({ _id: noteData._id }, { content: newContent })
             return
         }
         catch (err) {
